Added files via upload
We were never able to get the 60% speed up like the paper was expecting, but here is the code. We tried a lot of different things to hopefully improve it. We've also included the commands and results we've run so far on this updated version.
When using the simulator, if anyone wants to use a predictor, add -DBP # in the command line. 0 indicates no predictor, 1 indicates reference trace, and 2 indicates burst trace. If anyone plan on using the commands we're giving, make sure to change all the file locations. Also, if want to test the simulator, we suggest GCC with bank.i.